chore: Group the Progress Bar and Footer components into a single block.
